+++ Frends.Redis.GetValue/Frends.Redis.GetValue/Frends.Redis.GetValue.cs
@@ -1,8 +1,10 @@
-﻿namespace Frends.Redis.GetValue;
+﻿
+namespace Frends.Redis.GetValue;
 
 using System;
 using System.ComponentModel;
 using System.Threading.Tasks;
+using System.Linq;
 using Definitions;
 using Helpers;
 using StackExchange.Redis;
@@ -22,34 +24,48 @@ public static class Redis
     /// <param name="options">Exception settings.</param>
     /// <param name="connection">Connection info.</param>
     /// <returns>Object { string Value, bool Success, Error error }.</returns>
-    public static async Task<Result> GetValue([PropertyTab] Input input, [PropertyTab] Options options,
+    public static async Task<Result> GetValue(
+        [PropertyTab] Input input,
+        [PropertyTab] Options options,
         [PropertyTab] Connection connection)
     {
         try
         {
             if (string.IsNullOrEmpty(input.Key))
-                throw new ArgumentException("Key cannot be null or empty", nameof(input.Key));
+            {
+                throw new ArgumentException("Key cannot be null or empty", nameof(input));
+            }
+
             redis = await ConnectionMultiplexer.ConnectAsync(connection.ConnectionString);
             var db = redis.GetDatabase();
-            var value = await db.StringGetAsync(input.Key);
-            // TODO same ValueType needed + get values same as in tests for in SetValue task
-            // OR TRY THIS
-            // RedisType type = await db.KeyTypeAsync(key);
-            // return type switch
-            // {
-            //     RedisType.String => (string)await db.StringGetAsync(key),
-            //     RedisType.Hash => (await db.HashGetAllAsync(key)).ToDictionary(x => x.Name.ToString(),
-            //         x => x.Value.ToString()),
-            //     RedisType.List => (await db.ListRangeAsync(key)).Select(x => x.ToString()).ToList(),
-            //     RedisType.Set => (await db.SetMembersAsync(key)).Select(x => x.ToString()).ToList(),
-            //     RedisType.None => null, // Key does not exist
-            //     _ => throw new NotSupportedException($"Redis type {type} is not handled.")
-            // };
+            RedisType type = await db.KeyTypeAsync(input.Key);
 
+            var result = new Result { Success = true };
 
-            if (value.HasValue) return new Result(value.ToString());
+            switch (type)
+            {
+                case RedisType.String:
+                    result.StringValue = await db.StringGetAsync(input.Key);
+                    break;
+                case RedisType.List:
+                    var val = await db.ListRangeAsync(input.Key);
+                    result.ListValue = val.Select(x => x.ToString()).ToList();
+                    break;
+                case RedisType.Set:
+                    var setVal = await db.SetMembersAsync(input.Key);
+                    result.ListValue = setVal.Select(x => x.ToString()).ToList();
+                    break;
+                case RedisType.Hash:
+                    var hashVal = await db.HashGetAllAsync(input.Key);
+                    result.DictionaryValue = hashVal.ToDictionary(x => x.Name.ToString(), x => x.Value.ToString());
+                    break;
+                case RedisType.None:
+                    throw new Exception("Value not found");
+                default:
+                    throw new NotSupportedException($"Redis type {type} is not handled.");
+            }
 
-            throw new Exception("Value not found");
+            return result;
         }
         catch (Exception ex)
         {