+++ Frends.Redis.SetValue/Frends.Redis.SetValue.Tests/IntegrationTests.cs
@@ -33,32 +33,18 @@ public class IntegrationTests
 
         redis = await ConnectionMultiplexer.ConnectAsync(new ConfigurationOptions
         {
-            EndPoints =
-            {
-                connectionString,
-            },
-            AllowAdmin = true,
+            EndPoints = { connectionString, }, AllowAdmin = true,
         });
     }
 
     [SetUp]
     public void Setup()
     {
-        connection = new Connection
-        {
-            ConnectionString = connectionString,
-        };
+        connection = new Connection { ConnectionString = connectionString, };
 
-        input = new Input
-        {
-            Key = "test-key",
-            StringValue = "test-value",
-        };
+        input = new Input { Key = "test-key", StringValue = "test-value", };
 
-        options = new Options
-        {
-            ThrowErrorOnFailure = true,
-        };
+        options = new Options { ThrowErrorOnFailure = true, };
 
         var server = redis.GetServer(connectionString);
         server.FlushAllDatabases();
@@ -100,15 +86,7 @@ public class IntegrationTests
     {
         // Arrange
         var db = redis.GetDatabase();
-        var testVal = new Dictionary<string, string>
-        {
-            {
-                "Foo", "Bar"
-            },
-            {
-                "Moo", "Baz"
-            },
-        };
+        var testVal = new Dictionary<string, string> { { "Foo", "Bar" }, { "Moo", "Baz" }, };
         var entries = testVal.Select(x => new HashEntry(x.Key, x.Value)).ToArray();
         input.HashValue = testVal;
         input.ValueType = ValueType.Hash;
@@ -128,11 +106,7 @@ public class IntegrationTests
     {
         // Arrange
         var db = redis.GetDatabase();
-        var testVal = new List<string>
-        {
-            "foo",
-            "foo",
-        };
+        var testVal = new List<string> { "foo", "foo", };
         var entries = testVal.Select(x => (RedisValue)x).ToArray();
         input.ListValue = testVal;
         input.ValueType = ValueType.List;
@@ -152,11 +126,7 @@ public class IntegrationTests
     {
         // Arrange
         var db = redis.GetDatabase();
-        var testVal = new List<string>
-        {
-            "foo",
-            "foo",
-        };
+        var testVal = new List<string> { "foo", "foo", };
         var entries = testVal.Distinct().Select(x => (RedisValue)x).ToArray();
 
         input.ListValue = testVal;
@@ -167,6 +137,53 @@ public class IntegrationTests
 
         // Assert
         var savedValue = await db.SetMembersAsync(input.Key);
+        Assert.That(result.Success, Is.True);
+        Assert.IsNull(result.Error);
+        Assert.That(entries, Is.EquivalentTo(savedValue));
+    }
+
+    [Test]
+    public async Task AppendCollectionValueIsSuccessful()
+    {
+        // Arrange
+        var db = redis.GetDatabase();
+        var testVal = new List<string> { "foo", "foo" };
+        var expected = new List<string> { "foo", "foo", "foo", "foo" };
+        var entries = expected.Select(x => (RedisValue)x).ToArray();
+        input.ListValue = testVal;
+        input.ValueType = ValueType.List;
+        input.CollectionOperation = Operation.Append;
+        await Redis.SetValue(input, connection, options);
+
+        // Act
+        var result = await Redis.SetValue(input, connection, options);
+
+        // Assert
+        var savedValue = await db.ListRangeAsync(input.Key);
+        Assert.That(result.Success, Is.True);
+        Assert.IsNull(result.Error);
+        Assert.That(entries, Is.EquivalentTo(savedValue));
+    }
+
+    [Test]
+    public async Task OverwriteCollectionValueIsSuccessful()
+    {
+        // Arrange
+        var db = redis.GetDatabase();
+        var testVal = new List<string> { "foo1", "fo2" };
+        var expected = new List<string> { "foo3", "fo4" };
+        var entries = expected.Select(x => (RedisValue)x).ToArray();
+        input.ListValue = testVal;
+        input.ValueType = ValueType.List;
+        input.CollectionOperation = Operation.Overwrite;
+        await Redis.SetValue(input, connection, options);
+        input.ListValue = expected;
+
+        // Act
+        var result = await Redis.SetValue(input, connection, options);
+
+        // Assert
+        var savedValue = await db.ListRangeAsync(input.Key);
         Assert.That(result.Success, Is.True);
         Assert.IsNull(result.Error);
         Assert.That(entries, Is.EquivalentTo(savedValue));