+++ Frends.Redis.GetValue/CHANGELOG.md
@@ -1,9 +1,20 @@
 # Changelog
 
-## [1.1.0] - 2025-07-31
+## [2.0.0] - 2026-03-03
+
 ### Changed
+
+- Changed result values to distinguish between supported values (string, hash, set, list)
+- **[Breaking Change]** Renamed Result.Value to Result.StringValue to distinguish between other values
+
+## [1.1.0] - 2025-07-31
+
+### Changed
+
 - Make default error message empty by default
 
 ## [1.0.0] - 2025-05-13
+
 ### Changed
+
 - Initial implementation